backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pandas==2.1.3
numpy==1.26.2
requests==2.31.0
python-dotenv==1.0.0
crewai==0.11.0
langchain==0.0.350
transformers==4.35.2
torch>=2.2.0
pydantic==2.5.2
python-multipart==0.0.6
aiohttp==3.9.1
tqdm==4.66.1 